[WIP] Even more clippy-induced fixes and refactoring
