+++ src/cli/sgr_sequence/error.rs
