+++ src/core.rs
@@ -1,4 +1,5 @@
-pub(crate) mod exit_code;
-pub mod file_filtering;
-pub(crate) mod reader;
-pub mod request;
+pub mod context_accumulators;
+pub mod prospective_match_properties;
+pub mod reader;
+pub mod results_collection;
+pub mod top_bracket;
