--- src/matching_results.rs
@@ -1,4 +1,0 @@
-pub(crate) mod context_accumulators;
-pub mod result;
-pub(crate) mod result_collection;
-pub(crate) mod top_bracket;
