+++ tests/matching.rs
@@ -7,8 +7,7 @@
 )]
 #![expect(clippy::too_many_lines, reason = "It's tests, who cares?")]
 
-use fzgrep::{Targets, cli::args};
-use std::path::PathBuf;
+use fzgrep::cli;
 
 #[test]
 fn ascii_query() {
@@ -23,29 +22,17 @@ fn ascii_query() {
         "resources/tests/тест.txt",
         "resources/tests/测试.txt",
     ];
-    let request = args::make_request(cmd.into_iter().map(String::from));
-    assert_eq!(request.query, "contigous");
-    assert_eq!(
-        request.targets,
-        Targets::Files(vec![
-            PathBuf::from("resources/tests/👨‍🔬.txt"),
-            PathBuf::from("resources/tests/name with spaces.txt"),
-            PathBuf::from("resources/tests/test.txt"),
-            PathBuf::from("resources/tests/тест.txt"),
-            PathBuf::from("resources/tests/测试.txt")
-        ])
-    );
 
-    let results =
-        fzgrep::collect_all_matches(&request.query, &request.targets, &request.match_options)
-            .unwrap();
+    let request = cli::make_request(cmd.into_iter().map(String::from));
+    let results = fzgrep::collect_matches(&request.core).unwrap();
+
     assert_eq!(results.len(), 10);
 
     assert_eq!(
-        results[0].file_name.as_ref().unwrap(),
+        results[0].location.source_name.as_ref().unwrap(),
         "resources/tests/👨‍🔬.txt"
     );
-    assert_eq!(results[0].line_number.unwrap(), 6);
+    assert_eq!(results[0].location.line_number.unwrap(), 6);
     assert_eq!(results[0].matching_line, String::from("contiguous"));
     assert_eq!(results[0].fuzzy_match.score(), 116);
     assert_eq!(
@@ -54,10 +41,10 @@ fn ascii_query() {
     );
 
     assert_eq!(
-        results[1].file_name.as_ref().unwrap(),
+        results[1].location.source_name.as_ref().unwrap(),
         "resources/tests/name with spaces.txt"
     );
-    assert_eq!(results[1].line_number.unwrap(), 5);
+    assert_eq!(results[1].location.line_number.unwrap(), 5);
     assert_eq!(results[1].matching_line, String::from("contiguous"));
     assert_eq!(results[1].fuzzy_match.score(), 116);
     assert_eq!(
@@ -66,10 +53,10 @@ fn ascii_query() {
     );
 
     assert_eq!(
-        results[2].file_name.as_ref().unwrap(),
+        results[2].location.source_name.as_ref().unwrap(),
         "resources/tests/test.txt"
     );
-    assert_eq!(results[2].line_number.unwrap(), 2);
+    assert_eq!(results[2].location.line_number.unwrap(), 2);
     assert_eq!(results[2].matching_line, String::from("contiguous"));
     assert_eq!(results[2].fuzzy_match.score(), 116);
     assert_eq!(
@@ -78,10 +65,10 @@ fn ascii_query() {
     );
 
     assert_eq!(
-        results[3].file_name.as_ref().unwrap(),
+        results[3].location.source_name.as_ref().unwrap(),
         "resources/tests/тест.txt"
     );
-    assert_eq!(results[3].line_number.unwrap(), 5);
+    assert_eq!(results[3].location.line_number.unwrap(), 5);
     assert_eq!(results[3].matching_line, String::from("contiguous"));
     assert_eq!(results[3].fuzzy_match.score(), 116);
     assert_eq!(
@@ -90,10 +77,10 @@ fn ascii_query() {
     );
 
     assert_eq!(
-        results[4].file_name.as_ref().unwrap(),
+        results[4].location.source_name.as_ref().unwrap(),
         "resources/tests/测试.txt"
     );
-    assert_eq!(results[4].line_number.unwrap(), 3);
+    assert_eq!(results[4].location.line_number.unwrap(), 3);
     assert_eq!(results[4].matching_line, String::from("contiguous"));
     assert_eq!(results[4].fuzzy_match.score(), 116);
     assert_eq!(
@@ -102,10 +89,10 @@ fn ascii_query() {
     );
 
     assert_eq!(
-        results[5].file_name.as_ref().unwrap(),
+        results[5].location.source_name.as_ref().unwrap(),
         "resources/tests/👨‍🔬.txt"
     );
-    assert_eq!(results[5].line_number.unwrap(), 2);
+    assert_eq!(results[5].location.line_number.unwrap(), 2);
     assert_eq!(results[5].matching_line, String::from("Contiguous"));
     assert_eq!(results[5].fuzzy_match.score(), 115);
     assert_eq!(
@@ -114,10 +101,10 @@ fn ascii_query() {
     );
 
     assert_eq!(
-        results[6].file_name.as_ref().unwrap(),
+        results[6].location.source_name.as_ref().unwrap(),
         "resources/tests/name with spaces.txt"
     );
-    assert_eq!(results[6].line_number.unwrap(), 3);
+    assert_eq!(results[6].location.line_number.unwrap(), 3);
     assert_eq!(results[6].matching_line, String::from("Contiguous"));
     assert_eq!(results[6].fuzzy_match.score(), 115);
     assert_eq!(
@@ -126,10 +113,10 @@ fn ascii_query() {
     );
 
     assert_eq!(
-        results[7].file_name.as_ref().unwrap(),
+        results[7].location.source_name.as_ref().unwrap(),
         "resources/tests/test.txt"
     );
-    assert_eq!(results[7].line_number.unwrap(), 3);
+    assert_eq!(results[7].location.line_number.unwrap(), 3);
     assert_eq!(results[7].matching_line, String::from("Contiguous"));
     assert_eq!(results[7].fuzzy_match.score(), 115);
     assert_eq!(
@@ -138,10 +125,10 @@ fn ascii_query() {
     );
 
     assert_eq!(
-        results[8].file_name.as_ref().unwrap(),
+        results[8].location.source_name.as_ref().unwrap(),
         "resources/tests/тест.txt"
     );
-    assert_eq!(results[8].line_number.unwrap(), 6);
+    assert_eq!(results[8].location.line_number.unwrap(), 6);
     assert_eq!(results[8].matching_line, String::from("Contiguous"));
     assert_eq!(results[8].fuzzy_match.score(), 115);
     assert_eq!(
@@ -150,10 +137,10 @@ fn ascii_query() {
     );
 
     assert_eq!(
-        results[9].file_name.as_ref().unwrap(),
+        results[9].location.source_name.as_ref().unwrap(),
         "resources/tests/测试.txt"
     );
-    assert_eq!(results[9].line_number.unwrap(), 2);
+    assert_eq!(results[9].location.line_number.unwrap(), 2);
     assert_eq!(results[9].matching_line, String::from("Contiguous"));
     assert_eq!(results[9].fuzzy_match.score(), 115);
     assert_eq!(
@@ -175,65 +162,53 @@ fn emoji_query() {
         "resources/tests/тест.txt",
         "resources/tests/测试.txt",
     ];
-    let request = args::make_request(cmd.into_iter().map(String::from));
-    assert_eq!(request.query, "🐣🦀");
-    assert_eq!(
-        request.targets,
-        Targets::Files(vec![
-            PathBuf::from("resources/tests/name with spaces.txt"),
-            PathBuf::from("resources/tests/test.txt"),
-            PathBuf::from("resources/tests/👨‍🔬.txt"),
-            PathBuf::from("resources/tests/тест.txt"),
-            PathBuf::from("resources/tests/测试.txt")
-        ])
-    );
 
-    let results =
-        fzgrep::collect_all_matches(&request.query, &request.targets, &request.match_options)
-            .unwrap();
+    let request = cli::make_request(cmd.into_iter().map(String::from));
+    let results = fzgrep::collect_matches(&request.core).unwrap();
+
     assert_eq!(results.len(), 5);
 
     assert_eq!(
-        results[0].file_name.as_ref().unwrap(),
+        results[0].location.source_name.as_ref().unwrap(),
         "resources/tests/name with spaces.txt"
     );
-    assert_eq!(results[0].line_number.unwrap(), 1);
+    assert_eq!(results[0].location.line_number.unwrap(), 1);
     assert_eq!(results[0].matching_line, String::from("🐲🐣🐼🦀🦞🦠"));
     assert_eq!(results[0].fuzzy_match.score(), 4);
     assert_eq!(results[0].fuzzy_match.positions(), &vec![1, 3]);
 
     assert_eq!(
-        results[1].file_name.as_ref().unwrap(),
+        results[1].location.source_name.as_ref().unwrap(),
         "resources/tests/test.txt"
     );
-    assert_eq!(results[1].line_number.unwrap(), 6);
+    assert_eq!(results[1].location.line_number.unwrap(), 6);
     assert_eq!(results[1].matching_line, String::from("🐲🐣🐼🦀🦞🦠"));
     assert_eq!(results[1].fuzzy_match.score(), 4);
     assert_eq!(results[1].fuzzy_match.positions(), &vec![1, 3]);
 
     assert_eq!(
-        results[2].file_name.as_ref().unwrap(),
+        results[2].location.source_name.as_ref().unwrap(),
         "resources/tests/👨‍🔬.txt"
     );
-    assert_eq!(results[2].line_number.unwrap(), 5);
+    assert_eq!(results[2].location.line_number.unwrap(), 5);
     assert_eq!(results[2].matching_line, String::from("🐲🐣🐼🦀🦞🦠"));
     assert_eq!(results[2].fuzzy_match.score(), 4);
     assert_eq!(results[2].fuzzy_match.positions(), &vec![1, 3]);
 
     assert_eq!(
-        results[3].file_name.as_ref().unwrap(),
+        results[3].location.source_name.as_ref().unwrap(),
         "resources/tests/тест.txt"
     );
-    assert_eq!(results[3].line_number.unwrap(), 1);
+    assert_eq!(results[3].location.line_number.unwrap(), 1);
     assert_eq!(results[3].matching_line, String::from("🐲🐣🐼🦀🦞🦠"));
     assert_eq!(results[3].fuzzy_match.score(), 4);
     assert_eq!(results[3].fuzzy_match.positions(), &vec![1, 3]);
 
     assert_eq!(
-        results[4].file_name.as_ref().unwrap(),
+        results[4].location.source_name.as_ref().unwrap(),
         "resources/tests/测试.txt"
     );
-    assert_eq!(results[4].line_number.unwrap(), 4);
+    assert_eq!(results[4].location.line_number.unwrap(), 4);
     assert_eq!(results[4].matching_line, String::from("🐲🐣🐼🦀🦞🦠"));
     assert_eq!(results[4].fuzzy_match.score(), 4);
     assert_eq!(results[4].fuzzy_match.positions(), &vec![1, 3]);
@@ -252,110 +227,98 @@ fn cyrillic_query() {
         "resources/tests/👨‍🔬.txt",
         "resources/tests/测试.txt",
     ];
-    let request = args::make_request(cmd.into_iter().map(String::from));
-    assert_eq!(request.query, "тест");
-    assert_eq!(
-        request.targets,
-        Targets::Files(vec![
-            PathBuf::from("resources/tests/name with spaces.txt"),
-            PathBuf::from("resources/tests/test.txt"),
-            PathBuf::from("resources/tests/тест.txt"),
-            PathBuf::from("resources/tests/👨‍🔬.txt"),
-            PathBuf::from("resources/tests/测试.txt")
-        ])
-    );
 
-    let results =
-        fzgrep::collect_all_matches(&request.query, &request.targets, &request.match_options)
-            .unwrap();
+    let request = cli::make_request(cmd.into_iter().map(String::from));
+    let results = fzgrep::collect_matches(&request.core).unwrap();
+
     assert_eq!(results.len(), 10);
 
     assert_eq!(
-        results[0].file_name.as_ref().unwrap(),
+        results[0].location.source_name.as_ref().unwrap(),
         "resources/tests/name with spaces.txt"
     );
-    assert_eq!(results[0].line_number.unwrap(), 2);
+    assert_eq!(results[0].location.line_number.unwrap(), 2);
     assert_eq!(results[0].matching_line, String::from("тестування"));
     assert_eq!(results[0].fuzzy_match.score(), 46);
     assert_eq!(results[0].fuzzy_match.positions(), &vec![0, 1, 2, 3]);
 
     assert_eq!(
-        results[1].file_name.as_ref().unwrap(),
+        results[1].location.source_name.as_ref().unwrap(),
         "resources/tests/test.txt"
     );
-    assert_eq!(results[1].line_number.unwrap(), 5);
+    assert_eq!(results[1].location.line_number.unwrap(), 5);
     assert_eq!(results[1].matching_line, String::from("тестування"));
     assert_eq!(results[1].fuzzy_match.score(), 46);
     assert_eq!(results[1].fuzzy_match.positions(), &vec![0, 1, 2, 3]);
 
     assert_eq!(
-        results[2].file_name.as_ref().unwrap(),
+        results[2].location.source_name.as_ref().unwrap(),
         "resources/tests/тест.txt"
     );
-    assert_eq!(results[2].line_number.unwrap(), 4);
+    assert_eq!(results[2].location.line_number.unwrap(), 4);
     assert_eq!(results[2].matching_line, String::from("тестування"));
     assert_eq!(results[2].fuzzy_match.score(), 46);
     assert_eq!(results[2].fuzzy_match.positions(), &vec![0, 1, 2, 3]);
 
     assert_eq!(
-        results[3].file_name.as_ref().unwrap(),
+        results[3].location.source_name.as_ref().unwrap(),
         "resources/tests/👨‍🔬.txt"
     );
-    assert_eq!(results[3].line_number.unwrap(), 4);
+    assert_eq!(results[3].location.line_number.unwrap(), 4);
     assert_eq!(results[3].matching_line, String::from("тестування"));
     assert_eq!(results[3].fuzzy_match.score(), 46);
     assert_eq!(results[3].fuzzy_match.positions(), &vec![0, 1, 2, 3]);
 
     assert_eq!(
-        results[4].file_name.as_ref().unwrap(),
+        results[4].location.source_name.as_ref().unwrap(),
         "resources/tests/测试.txt"
     );
-    assert_eq!(results[4].line_number.unwrap(), 5);
+    assert_eq!(results[4].location.line_number.unwrap(), 5);
     assert_eq!(results[4].matching_line, String::from("тестування"));
     assert_eq!(results[4].fuzzy_match.score(), 46);
     assert_eq!(results[4].fuzzy_match.positions(), &vec![0, 1, 2, 3]);
 
     assert_eq!(
-        results[5].file_name.as_ref().unwrap(),
+        results[5].location.source_name.as_ref().unwrap(),
         "resources/tests/name with spaces.txt"
     );
-    assert_eq!(results[5].line_number.unwrap(), 4);
+    assert_eq!(results[5].location.line_number.unwrap(), 4);
     assert_eq!(results[5].matching_line, String::from("Текст"));
     assert_eq!(results[5].fuzzy_match.score(), 25);
     assert_eq!(results[5].fuzzy_match.positions(), &vec![0, 1, 3, 4]);
 
     assert_eq!(
-        results[6].file_name.as_ref().unwrap(),
+        results[6].location.source_name.as_ref().unwrap(),
         "resources/tests/test.txt"
     );
-    assert_eq!(results[6].line_number.unwrap(), 4);
+    assert_eq!(results[6].location.line_number.unwrap(), 4);
     assert_eq!(results[6].matching_line, String::from("Текст"));
     assert_eq!(results[6].fuzzy_match.score(), 25);
     assert_eq!(results[6].fuzzy_match.positions(), &vec![0, 1, 3, 4]);
 
     assert_eq!(
-        results[7].file_name.as_ref().unwrap(),
+        results[7].location.source_name.as_ref().unwrap(),
         "resources/tests/тест.txt"
     );
-    assert_eq!(results[7].line_number.unwrap(), 2);
+    assert_eq!(results[7].location.line_number.unwrap(), 2);
     assert_eq!(results[7].matching_line, String::from("Текст"));
     assert_eq!(results[7].fuzzy_match.score(), 25);
     assert_eq!(results[7].fuzzy_match.positions(), &vec![0, 1, 3, 4]);
 
     assert_eq!(
-        results[8].file_name.as_ref().unwrap(),
+        results[8].location.source_name.as_ref().unwrap(),
         "resources/tests/👨‍🔬.txt"
     );
-    assert_eq!(results[8].line_number.unwrap(), 1);
+    assert_eq!(results[8].location.line_number.unwrap(), 1);
     assert_eq!(results[8].matching_line, String::from("Текст"));
     assert_eq!(results[8].fuzzy_match.score(), 25);
     assert_eq!(results[8].fuzzy_match.positions(), &vec![0, 1, 3, 4]);
 
     assert_eq!(
-        results[9].file_name.as_ref().unwrap(),
+        results[9].location.source_name.as_ref().unwrap(),
         "resources/tests/测试.txt"
     );
-    assert_eq!(results[9].line_number.unwrap(), 6);
+    assert_eq!(results[9].location.line_number.unwrap(), 6);
     assert_eq!(results[9].matching_line, String::from("Текст"));
     assert_eq!(results[9].fuzzy_match.score(), 25);
     assert_eq!(results[9].fuzzy_match.positions(), &vec![0, 1, 3, 4]);
@@ -374,65 +337,53 @@ fn chinese_query() {
         "resources/tests/测试.txt",
         "resources/tests/👨‍🔬.txt",
     ];
-    let request = args::make_request(cmd.into_iter().map(String::from));
-    assert_eq!(request.query, "打电");
-    assert_eq!(
-        request.targets,
-        Targets::Files(vec![
-            PathBuf::from("resources/tests/name with spaces.txt"),
-            PathBuf::from("resources/tests/test.txt"),
-            PathBuf::from("resources/tests/тест.txt"),
-            PathBuf::from("resources/tests/测试.txt"),
-            PathBuf::from("resources/tests/👨‍🔬.txt"),
-        ])
-    );
 
-    let results =
-        fzgrep::collect_all_matches(&request.query, &request.targets, &request.match_options)
-            .unwrap();
+    let request = cli::make_request(cmd.into_iter().map(String::from));
+    let results = fzgrep::collect_matches(&request.core).unwrap();
+
     assert_eq!(results.len(), 5);
 
     assert_eq!(
-        results[0].file_name.as_ref().unwrap(),
+        results[0].location.source_name.as_ref().unwrap(),
         "resources/tests/name with spaces.txt"
     );
-    assert_eq!(results[0].line_number.unwrap(), 6);
+    assert_eq!(results[0].location.line_number.unwrap(), 6);
     assert_eq!(results[0].matching_line, String::from("打电动"));
     assert_eq!(results[0].fuzzy_match.score(), 17);
     assert_eq!(results[0].fuzzy_match.positions(), &vec![0, 1]);
 
     assert_eq!(
-        results[1].file_name.as_ref().unwrap(),
+        results[1].location.source_name.as_ref().unwrap(),
         "resources/tests/test.txt"
     );
-    assert_eq!(results[1].line_number.unwrap(), 1);
+    assert_eq!(results[1].location.line_number.unwrap(), 1);
     assert_eq!(results[1].matching_line, String::from("打电动"));
     assert_eq!(results[1].fuzzy_match.score(), 17);
     assert_eq!(results[1].fuzzy_match.positions(), &vec![0, 1]);
 
     assert_eq!(
-        results[2].file_name.as_ref().unwrap(),
+        results[2].location.source_name.as_ref().unwrap(),
         "resources/tests/тест.txt"
     );
-    assert_eq!(results[2].line_number.unwrap(), 3);
+    assert_eq!(results[2].location.line_number.unwrap(), 3);
     assert_eq!(results[2].matching_line, String::from("打电动"));
     assert_eq!(results[2].fuzzy_match.score(), 17);
     assert_eq!(results[2].fuzzy_match.positions(), &vec![0, 1]);
 
     assert_eq!(
-        results[3].file_name.as_ref().unwrap(),
+        results[3].location.source_name.as_ref().unwrap(),
         "resources/tests/测试.txt"
     );
-    assert_eq!(results[3].line_number.unwrap(), 1);
+    assert_eq!(results[3].location.line_number.unwrap(), 1);
     assert_eq!(results[3].matching_line, String::from("打电动"));
     assert_eq!(results[3].fuzzy_match.score(), 17);
     assert_eq!(results[3].fuzzy_match.positions(), &vec![0, 1]);
 
     assert_eq!(
-        results[4].file_name.as_ref().unwrap(),
+        results[4].location.source_name.as_ref().unwrap(),
         "resources/tests/👨‍🔬.txt"
     );
-    assert_eq!(results[4].line_number.unwrap(), 3);
+    assert_eq!(results[4].location.line_number.unwrap(), 3);
     assert_eq!(results[4].matching_line, String::from("打电动"));
     assert_eq!(results[4].fuzzy_match.score(), 17);
     assert_eq!(results[4].fuzzy_match.positions(), &vec![0, 1]);
